i hate gui design

+++ src/main/java/io/github/lijinhong11/supermines/listeners/BlockListener.java
@@ -61,7 +61,7 @@ public class BlockListener implements Listener {
         if (!treasures.isEmpty()) {
             for (Treasure treasure : treasures) {
                 if (treasure.getMatchedMaterials().contains(e.getBlock().getType())) {
-                    int chance = treasure.getChance();
+                    double chance = treasure.getChance();
                     if (NumberUtils.matchChance(chance)) {
                         world.dropItemNaturally(loc, treasure.getItemStack().clone());
                     }
